Improved 3D audio and MaxSources, replaced HashMap with Parallel HashMap library, added some objects and fixes, improved preloading

+++ Jazz2/nCine/Base/Iterator.h
@@ -120,7 +120,7 @@ namespace nCine
 		template <class RandomAccessIterator>
 		inline int distance(RandomAccessIterator& first, const RandomAccessIterator& last, RandomAccessIteratorTag)
 		{
-			return last - first;
+			return (int)(last - first);
 		}
 
 		/// Returns the distance in number of increments between two forward iterators
